+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 <a name="top"></a>
 
 Release notes
@@ -13,6 +14,8 @@ Releases marked with 🧪 were released on [the beta program](https://github.com
 Version 66
 ----------
 *in development*
+
+* 🔧 Trakt: show special error messages if account is locked or list limit exceeded.
 
 #### 66.0.5
 *2022-12-21*